feat: extracted logic into hook
