+++ src/hooks/Guilds/useIsProposalCreationAllowed/index.tsx
@@ -1,0 +1,29 @@
+import { useTypedParams } from 'Modules/Guilds/Hooks/useTypedParams';
+import { useMemo } from 'react';
+import { useAccount } from 'wagmi';
+import { useGuildConfig } from '../ether-swr/guild/useGuildConfig';
+import { useVotingPowerOf } from '../ether-swr/guild/useVotingPowerOf';
+
+const useIsProposalCreationAllowed = () => {
+  const { guildId } = useTypedParams();
+  const { address } = useAccount();
+  const { data: guildConfig } = useGuildConfig(guildId);
+  const { data: votingPower } = useVotingPowerOf({
+    contractAddress: guildId,
+    userAddress: address,
+  });
+
+  const isProposalCreationAllowed = useMemo(() => {
+    if (!guildConfig || !votingPower) {
+      return false;
+    }
+    if (votingPower.gte(guildConfig.votingPowerForProposalCreation)) {
+      return true;
+    }
+    return false;
+  }, [votingPower, guildConfig]);
+
+  return isProposalCreationAllowed;
+};
+
+export default useIsProposalCreationAllowed;
